+++ plugins/gs/src/apis/containerRegistry/index.ts
@@ -1,2 +1,3 @@
 export { ContainerRegistryClient } from './ContainerRegistryClient';
 export { containerRegistryApiRef } from './types';
+export type { TagInfo, TagsResponse, TagManifestResponse } from './types';