+++ plugins/gs/src/index.ts
@@ -40,5 +40,10 @@ export {
   isEntityHelmChartsAvailable as isEntityGSHelmChartsAvailable,
   isEntityInstallationResource as isEntityGSInstallationResource,
   isEntityKratixResource as isEntityGSKratixResource,
+  isEntityHelmChartTagged as isEntityGSVersionHistoryAvailable,
 } from './components/utils/entity';
+export { EntityVersionHistoryContent as EntityGSVersionHistoryContent } from './components/catalog/EntityVersionHistoryContent';
+export { EntityVersionHistoryCard as EntityGSVersionHistoryCard } from './components/catalog/EntityVersionHistoryCard';
+export { EntityReadmeCard as EntityGSReadmeCard } from './components/catalog/EntityReadmeCard';
+export { EntityChartProvider as EntityGSChartProvider } from './components/catalog/EntityChartContext';
 export { columnFactories as GSColumnFactories } from './components/catalog/columns';