+++ plugins/gs-backend/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @giantswarm/backstage-plugin-gs-backend
 
+## 0.3.0
+
+### Minor Changes
+
+- f665c62: Add Helm chart versions history.
+
 ## 0.2.0
 
 ### Minor Changes